+++ src/config/global-config.ts
@@ -15,7 +15,6 @@ export class GlobalConfigManager {
 
   static async load(context: vscode.ExtensionContext): Promise<GlobalConfig> {
     return context.globalState.get(this.CONFIG_KEY) ?? {
-      env: EnvironmentUtils.getSystemEnvironment(),
       manifestVersion: 0,
       isSetup: false,
       isSetupInProgress: false,
@@ -32,7 +31,6 @@ export class GlobalConfigManager {
     await context.globalState.update(this.CONFIG_KEY, undefined);
     // Load default config and fire event
     const defaultConfig: GlobalConfig = {
-      env: EnvironmentUtils.getSystemEnvironment(),
       manifestVersion: 0,
       isSetup: false,
       isSetupInProgress: false,