+++ src/types/config.ts
@@ -8,7 +8,6 @@ export interface GlobalConfig {
   isSetup: boolean;
   isSetupInProgress: boolean;
   manifestVersion: Number;
-  env: { [name: string]: string | undefined };
 }
 
 export interface ProbeInfo {